@@ -4,7 +4,7 @@
   
   <PropertyGroup>
     <TargetFramework>netstandard2.1</TargetFramework>
-    <LangVersion>9.0</LangVersion>
+    <LangVersion>10.0</LangVersion>
     <Nullable>enable</Nullable>
     <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
     <Authors>Tomáš Peťovský</Authors>
@@ -21,7 +21,7 @@
     <AssemblyVersion>3.0.0</AssemblyVersion>
     <FileVersion>3.0.0</FileVersion>
     <Version>3.0.0</Version>
-	<PackageVersion>3.0.0-alpha.4</PackageVersion>
+	<PackageVersion>3.0.0-alpha.5</PackageVersion>
   </PropertyGroup>
 
   <ItemGroup>
