@@ -1,40 +1,39 @@
 ﻿using Microsoft.AspNetCore.Authentication;
 using System;
 
-namespace NineDigit.NWS4.AspNetCore
+namespace NineDigit.NWS4.AspNetCore;
+
+public class NWS4AuthenticationSchemeOptions : AuthenticationSchemeOptions
 {
-    public class NWS4AuthenticationSchemeOptions : AuthenticationSchemeOptions
-    {
-        private TimeSpan _requestTimeWindow;
+    private TimeSpan _requestTimeWindow;
         
-        public NWS4AuthenticationSchemeOptions()
-        {
-            this.RequestTimeWindow = TimeSpan.FromSeconds(300);
-        }
-
-        public TimeSpan RequestTimeWindow
-        {
-            get => this._requestTimeWindow;
-            set
-            {
-                if (value.Ticks < 0)
-                    throw new ArgumentOutOfRangeException(nameof(value));
-
-                this._requestTimeWindow = value;
-            }
-        }
-
-        public string Realm { get; set; }
+    public NWS4AuthenticationSchemeOptions()
+    {
+        this.RequestTimeWindow = TimeSpan.FromSeconds(300);
     }
+
+    public TimeSpan RequestTimeWindow
+    {
+        get => this._requestTimeWindow;
+        set
+        {
+            if (value.Ticks < 0)
+                throw new ArgumentOutOfRangeException(nameof(value));
+
+            this._requestTimeWindow = value;
+        }
+    }
+
+    public string Realm { get; set; }
+}
     
-    public class NWS4AuthenticationSchemeOptions<TSignerOptions> : NWS4AuthenticationSchemeOptions
-        where TSignerOptions : SignerOptions, new()
+public class NWS4AuthenticationSchemeOptions<TSignerOptions> : NWS4AuthenticationSchemeOptions
+    where TSignerOptions : SignerOptions, new()
+{
+    public NWS4AuthenticationSchemeOptions()
     {
-        public NWS4AuthenticationSchemeOptions()
-        {
-            this.Signer = new TSignerOptions();
-        }
-        
-        public TSignerOptions Signer { get; }
+        this.Signer = new TSignerOptions();
     }
+        
+    public TSignerOptions Signer { get; }
 }
