@@ -1,22 +1,21 @@
 ﻿using Microsoft.AspNetCore.Authentication;
 using System;
 
-namespace NineDigit.NWS4.AspNetCore
+namespace NineDigit.NWS4.AspNetCore;
+
+internal class SystemClockDateTimeProvider : IDateTimeProvider, ISystemClock
 {
-    internal class SystemClockDateTimeProvider : IDateTimeProvider, ISystemClock
+    private readonly ISystemClock systemClock;
+
+    public SystemClockDateTimeProvider(ISystemClock systemClock)
     {
-        private readonly ISystemClock systemClock;
-
-        public SystemClockDateTimeProvider(ISystemClock systemClock)
-        {
-            this.systemClock = systemClock
-                ?? throw new ArgumentNullException(nameof(systemClock));
-        }
-
-        public DateTime UtcNow
-            => this.systemClock.UtcNow.UtcDateTime;
-
-        DateTimeOffset ISystemClock.UtcNow
-            => systemClock.UtcNow;
+        this.systemClock = systemClock
+                           ?? throw new ArgumentNullException(nameof(systemClock));
     }
+
+    public DateTime UtcNow
+        => this.systemClock.UtcNow.UtcDateTime;
+
+    DateTimeOffset ISystemClock.UtcNow
+        => systemClock.UtcNow;
 }
