@@ -1,24 +1,24 @@
 ﻿using System;
 using System.Runtime.Serialization;
 
-namespace NineDigit.NWS4
+namespace NineDigit.NWS4;
+
+[Serializable]
+public class SignatureExpiredException : Exception
 {
-    public class SignatureExpiredException : Exception
+    public SignatureExpiredException()
     {
-        public SignatureExpiredException()
-        {
-        }
+    }
 
-        public SignatureExpiredException(string message) : base(message)
-        {
-        }
+    public SignatureExpiredException(string message) : base(message)
+    {
+    }
 
-        public SignatureExpiredException(string message, Exception innerException) : base(message, innerException)
-        {
-        }
+    public SignatureExpiredException(string message, Exception innerException) : base(message, innerException)
+    {
+    }
 
-        protected SignatureExpiredException(SerializationInfo info, StreamingContext context) : base(info, context)
-        {
-        }
+    protected SignatureExpiredException(SerializationInfo info, StreamingContext context) : base(info, context)
+    {
     }
 }
