@@ -1,36 +1,35 @@
 ﻿using System.Text;
 
-namespace NineDigit.NWS4
+namespace NineDigit.NWS4;
+
+/// <summary>
+/// Various Http helper routines
+/// </summary>
+public static class HttpHelper
 {
     /// <summary>
-    /// Various Http helper routines
+    /// Helper routine to url encode canonicalized header names and values for safe
+    /// inclusion in the presigned url.
     /// </summary>
-    public static class HttpHelper
+    /// <param name="data">The string to encode</param>
+    /// <param name="isPath">Whether the string is a URL path or not</param>
+    /// <returns>The encoded string</returns>
+    public static string UrlEncode(string data, bool isPath = false)
     {
-        /// <summary>
-        /// Helper routine to url encode canonicalized header names and values for safe
-        /// inclusion in the presigned url.
-        /// </summary>
-        /// <param name="data">The string to encode</param>
-        /// <param name="isPath">Whether the string is a URL path or not</param>
-        /// <returns>The encoded string</returns>
-        public static string UrlEncode(string data, bool isPath = false)
+        // The Set of accepted and valid Url characters per RFC3986. Characters outside of this set will be encoded.
+        const string validUrlCharacters = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789-_.~";
+
+        var encoded = new StringBuilder(data.Length * 2);
+        string unreservedChars = string.Concat(validUrlCharacters, (isPath ? "/:" : ""));
+
+        foreach (char symbol in Encoding.UTF8.GetBytes(data))
         {
-            // The Set of accepted and valid Url characters per RFC3986. Characters outside of this set will be encoded.
-            const string validUrlCharacters = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789-_.~";
-
-            var encoded = new StringBuilder(data.Length * 2);
-            string unreservedChars = string.Concat(validUrlCharacters, (isPath ? "/:" : ""));
-
-            foreach (char symbol in Encoding.UTF8.GetBytes(data))
-            {
-                if (unreservedChars.IndexOf(symbol) != -1)
-                    encoded.Append(symbol);
-                else
-                    encoded.Append('%').Append(string.Format("{0:X2}", (int)symbol));
-            }
-
-            return encoded.ToString();
+            if (unreservedChars.IndexOf(symbol) != -1)
+                encoded.Append(symbol);
+            else
+                encoded.Append('%').Append(string.Format("{0:X2}", (int)symbol));
         }
+
+        return encoded.ToString();
     }
 }
