@@ -1,8 +1,7 @@
-﻿namespace NineDigit.NWS4
+﻿namespace NineDigit.NWS4;
+
+public interface IAuthDataSerializer
 {
-    public interface IAuthDataSerializer
-    {
-        public AuthData? Read(IHttpRequest request);
-        public void Write(IHttpRequest request, AuthData data);
-    }
+    public AuthData? Read(IHttpRequest request);
+    public void Write(IHttpRequest request, AuthData data);
 }
